Add files via upload
The Financial Fraud Prediction  project is a machine learning-based solution for detecting and preventing financial fraud. It uses various algorithms and techniques to analyze transactional data and identify patterns that may indicate fraud. The project may include pre-processing of financial data, feature engineering, model selection and evaluation, and deployment to a production environment. The goal of the project is to help financial institutions reduce the risk of fraud and improve their overall fraud detection capabilities.